Zooid and Receiver zipped Gerber files
Zipped Zooid and Receiver gerber files to make it easy to upload to OSH Park, PCB Way, Seeed, etc.

Issue with file on OSH Park for drill holes. Not sure if other PCB providers get the same errors when being uploaded.

When using the Zooid files to generate stencils not sure if bottom stencil needs to be mirrored for use with OSH Stencils solder paste stencil. Same for receiver.